Add files via upload
bit.io closed and is not accepting any new registration requests
I created a database on Amazon Web Services and linked to it using sequelize and MySqlWorkbench
The database configuration is is the config folder
the password is in the .env-file
Before you begin adding the cities to the database, please run the server.js; as soon as it runs, the server will connect to the database called "testapp"

Could you please see the work and advise what else needs to be done. I will make al corrections necessary. Thank you in advance
One more remark. Sometimes connection with the AWS does not work. I believe maybe it depends on the network protocol. I am still investigating